+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Breaking changes
+- The state of the extension is now saved in the reserved space by VS Code. `keep-context.json` will not be created.
+
 ## 0.1.2 - 2019-12-21
 
 ### Fixes